+++ CHANGELOG.md
@@ -9,7 +9,12 @@ Sections can be: Added Changed Deprecated Removed Fixed Security.
 ## [Unreleased]
 
 ### Fixed
-- Added Type field to PGN 130571.
+- #376: Added Type field to PGN 130571.
+- #377: Ensure PGNs are correct framing type (packet type) and fix PGN 130824 B&G Wind Data.
+- #365: Produce a `canboat.html` file
+
+### Changed
+- #366: PGN 127500 field lengths and framing type changed
 
 ## [4.10.0]